clean up namming, imports, and exports
